Updated the script
We're using here an internal API function for setting the root password.
It was generated by the same API function and the password sent to a
remote file called .lcpsql.
I removed the old "while read" with the mysqladmin command because it
requested root password in order to complete the setup.